add vendors theme

+++ Modules/Product/Http/Requests/UpdateProductRequest.php
@@ -23,7 +23,7 @@ class UpdateProductRequest extends FormRequest
             'product_quantity' => ['required', 'integer', 'min:1'],
             'product_cost' => ['required', 'numeric', 'max:2147483647'],
             'product_price' => ['required', 'numeric', 'max:2147483647'],
-            'product_stock_alert' => ['required', 'integer', 'min:0'],
+            'product_stock_alert' => ['required', 'numeric', 'min:0'],
             'product_order_tax' => ['nullable', 'numeric', 'min:0', 'max:100'],
             'product_tax_type' => ['nullable', 'numeric'],
             'product_note' => ['nullable', 'string', 'max:1000'],
